requirements.txt
requests
pygame
unicornhathd
python-dateutil
pytz
certifi
urllib3[secure]
pyOpenSSL
cryptography
idna
pillow
